backend/requirements.txt
aiosmtplib==5.1.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1

contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
fonttools==4.61.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.1
openai==1.3.7
packaging==25.0
passlib==1.7.4
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyparsing==3.3.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
gunicorn==21.2.0
qdrant-client
groq
google-generativeai
apscheduler
supabase>=2.3.4
